requirements.txt
apscheduler==3.10.4
asgiref==3.6.0
django==4.2.7
django-cors-headers==4.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
PyJWT==1.7.1
pytz==2020.5
setuptools==75.1.0
sqlparse==0.4.1
psycopg2-binary==2.9.9
python-dotenv==0.8.2
whitenoise==6.2.0